+++ src/components/inputs/BaseAutocomplete.vue
@@ -20,41 +20,55 @@
     :class="{ disableInput: disabled }"
   >
     <template v-slot:no-data>
-      <p id="no-resault">{{ $t(`autocomplete.${noResult || 'NoResult'}`) }}</p>
+      <p id="no-result">
+        {{ $t(`autocomplete.${validationFailedMsg ? validationFailedMsg : noResult || "NoResult"}`) }}
+      </p>
     </template>
   </v-autocomplete>
 </template>
 
 <script>
-import debounce from 'lodash/debounce';
+import debounce from "lodash/debounce";
+import { mapGetters } from "vuex";
 
 export default {
   data: () => ({
     item: null,
-    value: '',
+    value: "",
+    validationFailedMsg: null,
   }),
-  props: ['placeholder', 'items', 'background', 'icon', 'isLoading', 'minLength', 'noResult', 'disabled'],
+  computed: {
+    ...mapGetters(["error"]),
+  },
+  props: ["placeholder", "items", "background", "icon", "isLoading", "minLength", "noResult", "disabled", "validation"],
   methods: {
     onSelect() {
-      this.$emit('select', this.item);
+      this.$store.commit("setCurrentMailOrT", this.value);
+      this.$emit("select", this.item);
     },
     onEnter() {
-      this.item = '';
-      this.$emit('enter', this.value);
+      this.item = "";
+      this.$emit("enter", this.value);
     },
     onInput(value) {
       this.value = value;
-      this.onSearch(value);
+      const validationMsg = this.validation && value && value.length > 1 ? this.validation(value) : null;
+      if (!validationMsg) {
+        this.validationFailedMsg = false;
+        this.onSearch(value);
+      } else {
+        this.validationFailedMsg = validationMsg;
+      }
     },
     onSearch: debounce(function(value) {
-      if (typeof value === 'string' && value.length >= this.minLength) this.$emit('type', value);
+      if (typeof value === "string" && value.length >= this.minLength) this.$emit("type", value);
     }, 500),
   },
 };
 </script>
 
 <style scoped>
-#no-resault {
+#no-result {
   padding-right: 20px;
 }
 .disableInput {