+++ src/store/modules/auth.js
@@ -28,13 +28,11 @@ const actions = {
   },
   async addApproverInfos({ rootState, commit, dispatch }, user) {
     var approverInfos = {};
-
     await Promise.allSettled(
       rootState.configuration.externalNetworkDests.map(async (externalNetworkDest) => {
         const res = await usersApi.getApproverInfo(user.id, externalNetworkDest.value);
         approverInfos[externalNetworkDest.value] = res;
-      })).then((results) => 
-      {
+      })).then((results) => {
         if (results.some((result) => result.status === "rejected")) {
           // Retry after 1 minute
           setTimeout(() => dispatch("addApproverInfos", user), 10000);