+++ src/utils/networkDest.js
@@ -1,16 +1,16 @@
 import store from "@/store";
 
 export function getNetworkItemByDest(destination) {
-  const networkDest = store.state.configuration.externalNetworkDests.filter(
+  const networkDest = store.state.configuration.externalNetworkDests.find(
     (networkDest) => networkDest.value == destination
-  )[0];
+  );
   return networkDest;
 }
 
 export function getNetworkItemByAppId(appId) {
-  const networkDest = store.state.configuration.externalNetworkDests.filter(
+  const networkDest = store.state.configuration.externalNetworkDests.find(
     (networkDest) => networkDest.appID == appId
-  )[0];
+  );
   return networkDest;
 }